functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.0.0
google-generativeai>=0.3.0
Flask>=3.0.0

